Fix dtype mismatch in LinearEncoder.forward during evaluation
F.linear rejected float32 input against bfloat16 W_enc with:
  RuntimeError: expected mat1 and mat2 to have the same dtype,
                but got: float != c10::BFloat16

The crash surfaced during monosemanticity evaluation in the paper
runner after the linear CLT training completed. KANEncoder.forward()
already handled this correctly by casting input to the weight dtype
and casting the result back to the caller's dtype. Apply the same
pattern to LinearEncoder.

+++ spline_clt/linear_encoder.py
@@ -41,7 +41,7 @@ class LinearEncoder(nn.Module):
         Returns:
             Feature pre-activations of shape (..., n_features).
         """
-        return F.linear(x, self.W_enc)
+        return F.linear(x.to(self.W_enc.dtype), self.W_enc).to(x.dtype)
 
     def get_encoder_vectors(
         self, x: torch.Tensor, active_mask: torch.Tensor
